requirements.txt
pytest~=7.1.2
selene==2.0.0b4
allure-pytest